feat: abstract user registration logic and sign the user in after sucessfully registration
