+++ apps/mobile/src/hooks/use-user-register/index.ts
@@ -1,0 +1,1 @@
+export { default } from './use-user.register';
