+++ apps/mobile/src/navigation/types.ts
@@ -1,0 +1,7 @@
+import { UserType } from '@animavita/types';
+
+export type GetLocationScreenParamList = {
+  GetLocation: {
+    user: Pick<UserType, 'name' | 'email' | 'password'>;
+  };
+};
