+++ apps/mobile/src/services/sign-up.ts
@@ -1,7 +1,7 @@
-import { SignUpRequest } from '@animavita/types';
+import { SignUpRequest, SignUpResponse } from '@animavita/types';
 
 import client from './http-client';
 
 export const signUp = (user: SignUpRequest) => {
-  return client.post('/auth/signUp', user);
+  return client.post<SignUpResponse>('/auth/signUp', user);
 };
