+++ shared/types/dtos/auth/index.ts
@@ -1,2 +1,3 @@
 export * from "./sign-in-dto";
+export * from "./sign-up-dto";
 export * from "./auth-dto";
